song refactor
