+++ src/component/song/AlbumInfo.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function AlbumInfo() {
+  return (
+    <div>
+      album is
+    </div>
+  );
+}
