+++ src/component/song/ArtistInfo.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ArtistInfo() {
+  return (
+    <div>
+      artist=asdf
+    </div>
+  );
+}
