--- src/container/AlbumInfo.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function AlbumInfo() {
-  return (
-    <div>
-      album is
-    </div>
-  );
-}
