--- src/container/ArtistInfo.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function ArtistInfo() {
-  return (
-    <div>
-      artist=asdf
-    </div>
-  );
-}
